pyproject.toml
[tool.poetry]
name = "langchain-cookbook"
version = "0.1.0"
description = ""
authors = ["なるみ <toucans-cutouts0f@icloud.com>"]
readme = "README.md"
packages = [{include = "langchain_cookbook"}]

[tool.poetry.dependencies]
python = "^3.9"
langchain = "^0.0.235"
openai = "^0.27.8"
loguru = "^0.7.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
yapf = "^0.40.1"
toml = "^0.10.2"
flake8 = "^6.0.0"
pytest = "^7.4.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
